backend/requirements.txt
pillow
tensorflow
tensorflow-hub
fastapi
matplotlib
numpy
opencv-python
ipython
python-dotenv
pymongo
uvicorn
python-multipart